requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
asyncio==3.4.3
attrs==23.1.0
Babel==2.13.1
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
build==1.2.1
cachetools==5.4.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.5
chromadb==0.5.4
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.19.0
filelock==3.13.4
Flask==3.0.2
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
google-auth==2.32.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
groq==0.9.0
grpcio==1.65.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
humanfriendly==10.0
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.26.0
ipython==8.17.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter-tabnine==1.2.3
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.2
kubernetes==30.1.0
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-groq==0.1.6
langchain-text-splitters==0.2.2
langsmith==0.1.93
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
neo4j==5.22.0
nest-asyncio==1.5.8
networkx==3.3
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.6
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.0.0
playwright==1.45.0
posthog==3.5.0
prometheus-client==0.18.0
prompt-toolkit==3.0.41
protobuf==4.25.3
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
PyGetWindow==0.0.9
Pygments==2.16.1
PyMsgBox==1.0.9
pypdf==4.3.0
pyperclip==1.8.2
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytweening==1.2.0
pytz==2024.1
pywhatkit==5.4
# pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.13.0
rsa==4.9
selenium==4.17.2
Send2Trash==1.8.2
setuptools==71.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tenacity==8.5.0
terminado==0.18.0
tinycss2==1.2.1
tokenizers==0.19.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.3.3
tqdm==4.66.4
traitlets==5.13.0
trio==0.24.0
trio-websocket==0.11.1
typer==0.12.3
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.30.3
watchfiles==0.22.0
wcwidth==0.2.10
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
wheel==0.43.0
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
yarl==1.9.4
zipp==3.19.2